edit category and career models

+++ models/carrers.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Career struct {
-	CareerID    int          `gorm:"primaryKey;autoIncrement" json:"career_id"`
+	CareerID    int64        `gorm:"primaryKey;autoIncrement" json:"career_id"`
 	Name        string       `gorm:"not null" json:"name"`
 	Description string       `gorm:"default:NULL" json:"description"`
 	ShortDesc   string       `gorm:"default:NULL" json:"short_desc"`
@@ -13,7 +13,7 @@ type Career struct {
 }
 
 type Categories struct {
-	CategoryID int    `gorm:"primaryKey" json:"category_id"`
+	CategoryID int64  `gorm:"primaryKey" json:"category_id"`
 	Name       string `gorm:"not null" json:"name"`
 	ImageUrl   string `gorm:"default:NULL" json:"image_url"`
 }
